add populate script that backfill UTXOs used in mempool transactions

+++ migrations/sqls/20240323124855-init-up.sql
@@ -17,7 +17,7 @@ CREATE TABLE txs (
 CREATE TABLE txos (
   txid               TEXT            NOT NULL,
   idx                INT             NOT NULL,
-  amount             INT             NOT NULL,
+  amount             BIGINT          NOT NULL,
   spent_in_txid      TEXT            REFERENCES txs(txid),
   CONSTRAINT pk_txos_txid_idx PRIMARY KEY (txid, idx)
 ) WITH (OIDS=FALSE);
